requirements.txt
Flask
requests
python-dotenv
flask_sqlalchemy
psycopg2-binary
flask-login